pyproject.toml
[project]
name = "molescan-backend"
version = "0.1.0"
description = "Dermatology mole classification backend — CITDS 2026"
authors = [{ name = "Asfand Yar", email = "asfandyar@example.com" }]
requires-python = ">=3.11"

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "python-multipart>=0.0.9",
    "pillow>=10.3.0",
    "torch>=2.3.0",
    "torchvision>=0.18.0",
    "transformers>=4.41.0",
    "timm>=1.0.3",
    "numpy>=1.26.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
